big refactor to bring socket to dashboard scope and fix room bug
