+++ src/components/dashboard/StagingArea.jsx
@@ -4,11 +4,11 @@ import PostTrip from "./Post-Trip/PostTrip.jsx";
 import { Outlet } from "react-router-dom";
 
 
-export default function StagingArea({stops, addStop}) {
+export default function StagingArea({stops, addStop, messages, socket}) {
 
   return (
     <StagingAreaContainer>
-      <Outlet context={[stops, addStop]}/>
+      <Outlet context={[stops, addStop, messages, socket]}/>
     </StagingAreaContainer>
   );
 }
